requirements.txt
Django==2.1.5
pytz==2018.5
djangorestframework
pyyaml
requests